% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMatch.R
\name{betaMatch}
\alias{betaMatch}
\title{Beta matching for expert elicitations}
\usage{
betaMatch(par, elicited)
}
\arguments{
\item{par}{Target of optimisation; not set by user.}

\item{elicited}{Data frame of elicited values.}
}
\value{
No object is returned.
}
\description{
\code{betaMatch} This is the minimisation function for fitting a beta
distribution to expert elicited data using the IDEA protocol.
}

