% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicitator-package.R
\docType{package}
\name{elicitator-package}
\alias{elicitator}
\alias{elicitator-package}
\title{elicitator: An expert elicitation WEB app}
\description{
The goal of **elicitator** is to define an expert
    elicitation WEB app to answer with regard to _Case Study 2_ of EFSA's
    OC/EFSA/AMU/2019/02.
}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@unipd.it}

Authors:
\itemize{
  \item Danila Azzolina \email{danila.azzolina@unipd.it}
  \item Paola Berchialla \email{paola.berchialla@unito.it}
  \item Marco Ghidina \email{marco.ghidina@zetaresearch.com}
  \item Dario Gregori \email{dario.gregori@unipd.it}
}

}
\keyword{internal}
