% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicitatoR.R
\name{elicitatoR}
\alias{elicitatoR}
\title{Run the elicitatoR}
\usage{
elicitatoR()
}
\value{
No objects are returned. Elicitations are stored in a database inside
    the data folder. Both the database and data folder will be created inside
    the working directory if they don't exist.
}
\description{
\code{elicitatoR} Starts the elicitatoR shiny app.
}
\examples{
\dontrun{
elicitatoR()
}

}

