% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usr.R
\name{get_usr}
\alias{get_usr}
\title{Get user info}
\usage{
get_usr(usr, psw)
}
\arguments{
\item{usr}{(chr) username}

\item{psw}{(chr) password}
}
\value{
a \link[tibble:tibble-package]{tibble} including the following
user info from the \strong{users} database (see the section Notes):
\itemize{
\item \code{id_users}: the user's ID
\item \code{username}: the user's username
\item \code{name}: the user's name
\item \code{lastname}: the user's surname
\item \code{role}: the user's role group (1 = expert elicitator, 2 = admin)
}
}
\description{
Extract the user information from the \strong{elicitator} MySQL database
stored on the server, and return them like a
\link[tibble:tibble-package]{tibble}.
}
\note{
to access the database the following environmental variable must
be set-up:\preformatted{  - `USER_TEST`: username to access the **elicitator** MySQL
      database on the server
  - `PSW_TEST`: password to access **elicitator**
  - `TBL_KEY`: AES key used to encrypt the passwords on the
    **users** table of **elicitator**
}
}
\examples{
\dontrun{
    # On r-ubesp server
    psw <- Sys.getenv("USR_PSW")

    get_usr("user_a", psw)
    get_usr("user_a", psw)
    get_usr("admin", psw)
}
}
