% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_stored_vals.R
\name{last_stored_vals}
\alias{last_stored_vals}
\title{Read last expert elicitated stored values}
\usage{
last_stored_vals(usr, psw)
}
\arguments{
\item{usr}{(chr) username}

\item{psw}{(chr) password}
}
\value{
a \link[tibble:tibble-package]{tibble} including the following
informations from the \strong{opinions} database (see the section Notes):\preformatted{- `users_id`: the id of the user
- `date_opinion`: the date in which the opinion was stored
- `perc1`:  the 1th percentile elicitated
- `perc25`: the 25th percentile elicitated
- `perc50`: the 50th percentile elicitated
- `perc75`: the 75th percentile elicitated
- `perc99`: the 99th percentile elicitated
}
}
\description{
Read the last stored elicitated values on the \strong{opinions} table of
the \strong{elicitator} MySQL database stored on the server.
}
\note{
to access the database the following environmental variable must
be set-up:\preformatted{  - `USER_TEST`: username to access the **elicitator** MySQL
      database on the server
  - `PSW_TEST`: password to access **elicitator**
  - `TBL_KEY`: AES key used to encrypt the passwords on the
    **users** table of **elicitator**
}
}
\examples{
\dontrun{
    # On r-ubesp server
    psw <- Sys.getenv("USR_PSW")

    last_stored_vals("user_a", psw)
}
}
