% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupData.R
\name{setupData}
\alias{setupData}
\title{Setup a new database}
\usage{
setupData(dbName)
}
\arguments{
\item{dbName}{The name of the database to store the elicitations in.}
}
\value{
No object is returned, the database is created. Note: the database is
    created in the data directory (which is created if it doesn't exist).
}
\description{
\code{setupData} Sets up a new database to store the expert elicitations in.
}

