% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_expert_vals.R
\name{store_expert_vals}
\alias{store_expert_vals}
\title{Store expert elicitated values}
\usage{
store_expert_vals(usr, psw, vals)
}
\arguments{
\item{usr}{(chr) username}

\item{psw}{(chr) password}

\item{vals}{(dbl) numerical vector of length five, containing, in
order, the elicitated values for the 1st, 25th, 50th, 75th, 99th
percentiles}
}
\value{
invisibly \code{TRUE}
}
\description{
Writes the elicitated values on the \strong{opinions} table of the
\strong{elicitator} MySQL database stored on the server.
}
\note{
to access the database the following environmental variable must
be set-up:\preformatted{  - `USER_TEST`: username to access the **elicitator** MySQL
      database on the server
  - `PSW_TEST`: password to access **elicitator**
  - `TBL_KEY`: AES key used to encrypt the passwords on the
    **users** table of **elicitator**
}
}
\examples{
\dontrun{
    # On r-ubesp server
    psw <- Sys.getenv("USR_PSW")

    store_expert_vals("user_a", psw, 1:5)
}
}
